# backend/postman/openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI Documentation","description":"\nAPI endpoints for LLDP project <https://www.lldp.duckdns.org>\n\n\n---\n\n**Contact**\n\n\n* Author: Matias Jose Varrone\n\n* Email: mativarrone2@gmail.com\n\nMarch, 2022\n\n---\n\n","license":{"name":"nginx 1.21.6","url":"https://nginx.org/LICENSE"},"version":"0.0.1"},"paths":{"/token":{"post":{"tags":["Login"],"summary":"Login For Access Token","operationId":"login_for_access_token_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/me/":{"get":{"tags":["Login"],"summary":"Read Users Me","operationId":"read_users_me_users_me__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/users/me/items/":{"get":{"tags":["Login"],"summary":"Read Own Items","operationId":"read_own_items_users_me_items__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/":{"get":{"tags":["Root"],"summary":"Root","description":"Used when click on Home tab -->\nFile: Home.vue, Section: mounted()","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/diagram":{"get":{"tags":["Diagram"],"summary":"Diagram","description":"Used when click on Diagram tab -->\nFile: components/NetworkDiagram.vue, Section: mounted()","operationId":"diagram_diagram_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/device_type_list":{"get":{"tags":["Lists"],"summary":"Get Device Type List","description":"Used when click on Add Tab -->\nFile: AddDevice.vue, Section: mounted()","operationId":"get_device_type_list_device_type_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/logs":{"get":{"tags":["Logs"],"summary":"Get Logs","description":"Used when click on Logs tab -->\nFile: Logs.vue, Section: mounted()","operationId":"get_logs_logs_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/add_device/{hostname}":{"post":{"tags":["Devices"],"summary":"Add Device","description":"Used when click on Add Button on Add Tab -->\nFile: AddDevice.vue, Section: method: add_device_button()","operationId":"add_device_add_device__hostname__post","parameters":[{"required":true,"schema":{"title":"Hostname","type":"string"},"name":"hostname","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddDevice"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/modify_device/{hostname}":{"put":{"tags":["Devices"],"summary":"Modify Device","description":"Used when click on Modify Button on Modify Tab -->\nFile: ModifyDevice.vue, Section: method: modify_device_button()","operationId":"modify_device_modify_device__hostname__put","parameters":[{"required":true,"schema":{"title":"Hostname","type":"string"},"name":"hostname","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDevice"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/delete_device/{hostname}":{"delete":{"tags":["Devices"],"summary":"Delete Device","description":"Used when click on Delete Button on Delete Tab -->\nFile: DeleteDevice.vue, Section: method: delete_device_button()","operationId":"delete_device_delete_device__hostname__delete","parameters":[{"required":true,"schema":{"title":"Hostname","type":"string"},"name":"hostname","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteDevice"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/hostname_list":{"get":{"tags":["Lists"],"summary":"Get Hostname List","description":"Used when click on Delete Tab -->\nFile: DeleteDevice.vue, Section: mounted()\n\nAlso, after Delete button on Delete Tab is pressed -->\nFile: DeleteDevice.vue, Section: method: update_info_after_button_clicked()\n\nAlso, after Modify button on Modify Tab is pressed -->\nFile: ModifyDevice.vue, Section: method: update_info_after_modify_button_clicked()","operationId":"get_hostname_list_hostname_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/inventory_list":{"get":{"tags":["Lists"],"summary":"Get Inventory List","description":"Used when click on Inventory Tab -->\nFile: Inventory.vue, Section: mounted()","operationId":"get_inventory_list_inventory_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/about":{"get":{"tags":["About"],"summary":"About","description":"Used when click on About tab -->\nFile: About.vue, Section: mounted()","operationId":"about_about_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/device_values_to_modify/{hostname}":{"get":{"tags":["Lists"],"summary":"Get Device Values To Modify","description":"Used when click on an item from selectable list on Modify Tab -->\nFile: ModifyDevice.vue, Section: method: switchSelect(event)\n\nAlso, update_info_after_modify_button_clicked uses this endpoint.\nupdate_info_after_modify_button_clicked is only executed when there are no\nerrors when Modify Button is clicked -->\nFile: ModifyDevice.vue, Section: method: update_info_after_modify_button_clicked()","operationId":"get_device_values_to_modify_device_values_to_modify__hostname__get","parameters":[{"required":true,"schema":{"title":"Hostname","type":"string"},"name":"hostname","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/dashboard":{"get":{"tags":["Root"],"summary":"Dashboard","description":"Called when user logged in successfully -->\nFile: Dashboard.vue, Section: mounted()","operationId":"dashboard_dashboard_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/logout":{"get":{"tags":["Login"],"summary":"Logout","description":"Called when user click in Log out Tab -->\nFile: Logout.vue, Section: mounted()","operationId":"logout_logout_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/get_diagram_name_list":{"get":{"tags":["History"],"summary":"Get Name List","description":"Used when click on History tab -->\nFile: Todo-History.vue, Section: mounted()","operationId":"get_name_list_get_diagram_name_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/get_diagram_info_by_id/{id}":{"get":{"tags":["History"],"summary":"Get Diagram Info Id","description":"Used when click on any element on History Tab list -->\nFile: Todo-History.vue, Section: methods: clickRow(id)","operationId":"get_diagram_info_id_get_diagram_info_by_id__id__get","parameters":[{"required":true,"schema":{"title":"Id","type":"integer"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/get_permission_admin_value":{"get":{"tags":["Permissions"],"summary":"Get Permission","description":"Used by every endpoint to check if user has admin permissions (admin_permissions = 1)\nor not (admin_permissions = 0). \n\nThen, this value is used to show an appropiate Header: Todo-Header.vue or Todo-HeaderNoAdmin.vue","operationId":"get_permission_get_permission_admin_value_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"AddDevice":{"title":"AddDevice","required":["device_type","hostname","port","username","password","secret","conn_timeout"],"type":"object","properties":{"device_type":{"title":"Device Type","type":"string"},"hostname":{"title":"Hostname","type":"string"},"port":{"title":"Port","type":"integer"},"username":{"title":"Username","type":"string"},"password":{"title":"Password","type":"string"},"secret":{"title":"Secret","type":"string"},"conn_timeout":{"title":"Conn Timeout","type":"integer"}}},"Body_login_for_access_token_token_post":{"title":"Body_login_for_access_token_token_post","required":["username","password"],"type":"object","properties":{"grant_type":{"title":"Grant Type","pattern":"password","type":"string"},"username":{"title":"Username","type":"string"},"password":{"title":"Password","type":"string"},"scope":{"title":"Scope","type":"string","default":""},"client_id":{"title":"Client Id","type":"string"},"client_secret":{"title":"Client Secret","type":"string"}}},"DeleteDevice":{"title":"DeleteDevice","required":["hostname"],"type":"object","properties":{"hostname":{"title":"Hostname","type":"string"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Token":{"title":"Token","required":["access_token","token_type"],"type":"object","properties":{"access_token":{"title":"Access Token","type":"string"},"token_type":{"title":"Token Type","type":"string"}}},"UpdateDevice":{"title":"UpdateDevice","required":["device_type","hostname","port","username","password","secret","conn_timeout"],"type":"object","properties":{"device_type":{"title":"Device Type","type":"string"},"hostname":{"title":"Hostname","type":"string"},"port":{"title":"Port","type":"integer"},"username":{"title":"Username","type":"string"},"password":{"title":"Password","type":"string"},"secret":{"title":"Secret","type":"string"},"conn_timeout":{"title":"Conn Timeout","type":"integer"}}},"User":{"title":"User","required":["username"],"type":"object","properties":{"username":{"title":"Username","type":"string"},"email":{"title":"Email","type":"string"},"full_name":{"title":"Full Name","type":"string"},"disabled":{"title":"Disabled","type":"boolean"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"token"}}}}},"tags":[{"name":"Login","description":"Endpoints related to login users"},{"name":"Root"},{"name":"Diagram","description":"Connect to devices in db, get **LLDP** data and return it to the frontend"},{"name":"Devices","description":"Perform **Actions** on _devices_"},{"name":"Lists"},{"name":"Logs"},{"name":"About"},{"name":"History"},{"name":"Permissions","description":"Related to Header display"}]}